requirements.txt
requests
google-cloud-dns